patch 9.0.1849: CI error on different signedness in ex_cmds.c
Problem:  CI error on different signedness
Solution: cast unsigned to int

+++ src/ex_cmds.c
@@ -4651,7 +4651,7 @@ ex_substitute(exarg_T *eap)
 		mch_memmove(new_end, sub_firstline + copycol, (size_t)copy_len);
 		new_end += copy_len;
 
-		if (new_start_len - copy_len < sublen)
+		if ((int)new_start_len - copy_len < sublen)
 		    sublen = new_start_len - copy_len - 1;
 
 #ifdef FEAT_EVAL
